requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-solo==2.0.0
django-streamfield==1.4.4
filelock==3.8.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
six==1.16.0
sqlparse==0.4.3
tox==3.27.0
virtualenv==20.16.6
